@@ -5,23 +5,33 @@
   "author": "",
   "license": "ISC",
   "engines": {
-    "node": ">= 12.0.0"
+    "node": ">= 18.0.0"
   },
+  "type": "module",
   "scripts": {
-    "test": "jest",
-    "start": "PORT_SERVER=3000 node src/server",
-    "lint": "eslint src",
+    "test": "npm run test:v",
+    "test:v": "vitest run",
+    "test:j": "NODE_OPTIONS=--experimental-vm-modules jest",
+    "start": "PORT_SERVER=3000 node src/start",
+    "lint": "eslint src test",
     "lint:fix": "npm run lint -- --fix"
   },
-  "dependencies": {},
   "devDependencies": {
-    "eslint-plugin-jsdoc": "37.7.0",
+    "@types/jest": "29.5.10",
+    "@types/node": "20.9.4",
+    "eslint": "8.54.0",
+    "eslint-config-standard": "17.1.0",
+    "eslint-plugin-import": "2.29.0",
+    "eslint-plugin-jsdoc": "46.9.0",
+    "eslint-plugin-n": "16.3.1",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "6.1.1",
     "husky": "4.2.5",
-    "lint-staged": "12.3.2",
-    "jest": "27.4.7",
-    "jest-extended": "2.0.0",
-    "standard": "16.0.4",
-    "supertest": "6.2.0"
+    "jest": "29.7.0",
+    "jest-extended": "4.0.2",
+    "lint-staged": "15.1.0",
+    "supertest": "6.3.3",
+    "vitest": "^0.34.6"
   },
   "husky": {
     "hooks": {
