@@ -1,14 +1,12 @@
-// @ts-check
-
-const { Server } = require('./server')
+import { WebServer } from './server.js'
 
 const hostname = '127.0.0.1'
 const port = process.env.PORT_SERVER || '3000'
 
-Server()
+WebServer({})
 	.listen(Number(port), hostname, () => {
 		console.log(`Server running at http://${hostname}:${port}/`)
 	})
 	.on('error', err => {
-		console.error(err)
+		console.error('Cath error during initalization server:', err)
 	})
